# gateway/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 百舸API网关
  description: |
    统一百舸各模块API规范和数据结构的中间层网关服务。

    所有接口采用RPC风格，通过 `?action=XXX` 参数区分操作。
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000/v1
    description: 本地开发服务器
...
paths:
  /aihc/jobs/DescribeJobs:
    get:
      tags:
        - Jobs
      summary: Jobs - DescribeJobs
      description: 查询指定资源池和队列下的训练任务列表
      operationId: describeJobs
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: resourcePoolId
          in: query
          required: true
          schema:
            type: string
          description: 资源池ID
        - name: queue
          in: query
          schema:
            type: string
          description: 队列名称
        - name: status
          in: query
          schema:
            type: string
          description: 任务状态过滤
        - name: keyword
          in: query
          schema:
            type: string
          description: 搜索关键字
        - name: pageNumber
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
  /aihc/jobs/DescribeJob:
    get:
      tags:
        - Jobs
      summary: Jobs - DescribeJob
      description: 查询指定资源池和队列下的训练任务列表
      operationId: describeJob
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: resourcePoolId
          in: query
          required: true
          schema:
            type: string
          description: 资源池ID
        - name: queue
          in: query
          schema:
            type: string
          description: 队列名称
        - name: jobId
          in: query
          schema:
            type: string
          description: 任务ID（DescribeJob时必填）
        - name: needDetail
          in: query
          schema:
            type: boolean
          description: 是否需要详情（DescribeJob时可选）
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
  /aihc/jobs/CreateJob:
    post:
      tags:
        - Jobs
      summary: Jobs - CreateJob
      description: 创建或删除训练任务
      operationId: createJob
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateJobResponse"
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateJobResponse"
  /aihc/jobs/DeleteJob:
    post:
      tags:
        - Jobs
      summary: Jobs - DeleteJob
      description: 创建或删除训练任务
      operationId: deleteJob
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteJobRequest"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteJobResponse"
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateJobResponse"
  /aihc/services/DescribeServices:
    get:
      tags:
        - Services
      summary: Services - DescribeServices
      operationId: describeServices
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: serviceId
          in: query
          schema:
            type: string
          description: 服务ID（DescribeService时必填）
        - name: pageNumber
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceListResponse"
  /aihc/services/DescribeService:
    get:
      tags:
        - Services
      summary: Services - DescribeService
      operationId: describeService
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: serviceId
          in: query
          schema:
            type: string
          description: 服务ID（DescribeService时必填）
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceDetail"
  /aihc/services/CreateService:
    post:
      tags:
        - Services
      summary: Services - CreateService
      operationId: createService
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "200":
          description: 成功
        "201":
          description: 创建成功
  /aihc/services/DeleteService:
    post:
      tags:
        - Services
      summary: Services - DeleteService
      operationId: deleteService
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteServiceRequest"
      responses:
        "200":
          description: 成功
        "201":
          description: 创建成功
  /aihc/dev-instances/DescribeDevInstances:
    get:
      tags:
        - DevInstances
      summary: DevInstances - DescribeDevInstances
      operationId: describeDevInstances
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: devInstanceId
          in: query
          schema:
            type: string
          description: 开发机ID（DescribeDevInstance时必填）
        - name: resourcePoolId
          in: query
          schema:
            type: string
        - name: queueName
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: pageNumber
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DevInstanceListResponse"
  /aihc/dev-instances/DescribeDevInstance:
    get:
      tags:
        - DevInstances
      summary: DevInstances - DescribeDevInstance
      operationId: describeDevInstance
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: devInstanceId
          in: query
          schema:
            type: string
          description: 开发机ID（DescribeDevInstance时必填）
        - name: resourcePoolId
          in: query
          schema:
            type: string
        - name: queueName
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DevInstanceDetail"
  /aihc/dev-instances/CreateDevInstance:
    post:
      tags:
        - DevInstances
      summary: DevInstances - CreateDevInstance
      operationId: createDevInstance
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDevInstanceRequest"
      responses:
        "200":
          description: 成功
        "201":
          description: 创建成功
  /aihc/dev-instances/DeleteDevInstance:
    post:
      tags:
        - DevInstances
      summary: DevInstances - DeleteDevInstance
      operationId: deleteDevInstance
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteDevInstanceRequest"
      responses:
        "200":
          description: 成功
        "201":
          description: 创建成功
  /aihc/datasets/DescribeDatasets:
    get:
      tags:
        - Datasets
      summary: Datasets - DescribeDatasets
      operationId: describeDatasets
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: datasetId
          in: query
          schema:
            type: string
          description: 数据集ID
        - name: keyword
          in: query
          schema:
            type: string
        - name: storageType
          in: query
          schema:
            type: string
            enum:
              - pfs
              - bos
        - name: pageNumber
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetListResponse"
  /aihc/datasets/DescribeDataset:
    get:
      tags:
        - Datasets
      summary: Datasets - DescribeDataset
      operationId: describeDataset
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: datasetId
          in: query
          schema:
            type: string
          description: 数据集ID
        - name: keyword
          in: query
          schema:
            type: string
        - name: storageType
          in: query
          schema:
            type: string
            enum:
              - pfs
              - bos
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetDetail"
  /aihc/datasets/DescribeDatasetVersions:
    get:
      tags:
        - Datasets
      summary: Datasets - DescribeDatasetVersions
      operationId: describeDatasetVersions
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: datasetId
          in: query
          schema:
            type: string
          description: 数据集ID
        - name: keyword
          in: query
          schema:
            type: string
        - name: storageType
          in: query
          schema:
            type: string
            enum:
              - pfs
              - bos
        - name: pageNumber
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetVersionListResponse"
  /aihc/datasets/CreateDataset:
    post:
      tags:
        - Datasets
      summary: Datasets - CreateDataset
      operationId: createDataset
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDatasetRequest"
      responses:
        "200":
          description: 成功
        "201":
          description: 创建成功
...
  /aihc/datasets/DeleteDataset:
    post:
      tags:
        - Datasets
      summary: Datasets - DeleteDataset
      operationId: deleteDataset
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteDatasetRequest"
      responses:
        "200":
          description: 成功
        "201":
          description: 创建成功
...
  /aihc/models/DescribeModels:
    get:
      tags:
        - Models
      summary: Models - DescribeModels
      operationId: describeModels
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: modelId
          in: query
          schema:
            type: string
          description: 模型ID
        - name: keyword
          in: query
          schema:
            type: string
        - name: pageNumber
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelListResponse"
  /aihc/models/DescribeModel:
    get:
      tags:
        - Models
      summary: Models - DescribeModel
      operationId: describeModel
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: modelId
          in: query
          schema:
            type: string
          description: 模型ID
        - name: keyword
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelDetail"
  /aihc/models/DescribeModelVersions:
    get:
      tags:
        - Models
      summary: Models - DescribeModelVersions
      operationId: describeModelVersions
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: modelId
          in: query
          schema:
            type: string
          description: 模型ID
        - name: keyword
          in: query
          schema:
            type: string
        - name: pageNumber
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelVersionListResponse"
  /aihc/models/CreateModel:
    post:
      tags:
        - Models
      summary: Models - CreateModel
      operationId: createModel
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModelRequest"
      responses:
        "200":
          description: 成功
        "201":
          description: 创建成功
...
  /aihc/models/DeleteModel:
    post:
      tags:
        - Models
      summary: Models - DeleteModel
      operationId: deleteModel
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteModelRequest"
      responses:
        "200":
          description: 成功
        "201":
          description: 创建成功
...
  /aihc/resource-pools/DescribeResourcePools:
    get:
      tags:
        - ResourcePools
      summary: ResourcePools - DescribeResourcePools
      operationId: describeResourcePools
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: resourcePoolId
          in: query
          schema:
            type: string
          description: 资源池ID
        - name: queueId
          in: query
          schema:
            type: string
          description: 队列ID
        - name: nodeName
          in: query
          schema:
            type: string
          description: 节点名称
        - name: type
          in: query
          schema:
            type: string
            enum:
              - selfManaged
              - managed
            default: selfManaged
        - name: keyword
          in: query
          schema:
            type: string
        - name: pageNumber
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourcePoolListResponse"
  /aihc/resource-pools/DescribeResourcePool:
    get:
      tags:
        - ResourcePools
      summary: ResourcePools - DescribeResourcePool
      operationId: describeResourcePool
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: resourcePoolId
          in: query
          schema:
            type: string
          description: 资源池ID
        - name: queueId
          in: query
          schema:
            type: string
          description: 队列ID
        - name: nodeName
          in: query
          schema:
            type: string
          description: 节点名称
        - name: type
          in: query
          schema:
            type: string
            enum:
              - selfManaged
              - managed
            default: selfManaged
        - name: keyword
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourcePoolDetail"
  /aihc/resource-pools/DescribeQueues:
    get:
      tags:
        - ResourcePools
      summary: ResourcePools - DescribeQueues
      operationId: describeQueues
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: resourcePoolId
          in: query
          schema:
            type: string
          description: 资源池ID
        - name: queueId
          in: query
          schema:
            type: string
          description: 队列ID
        - name: nodeName
          in: query
          schema:
            type: string
          description: 节点名称
        - name: type
          in: query
          schema:
            type: string
            enum:
              - selfManaged
              - managed
            default: selfManaged
        - name: keyword
          in: query
          schema:
            type: string
        - name: pageNumber
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueListResponse"
  /aihc/resource-pools/DescribeQueue:
    get:
      tags:
        - ResourcePools
      summary: ResourcePools - DescribeQueue
      operationId: describeQueue
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: resourcePoolId
          in: query
          schema:
            type: string
          description: 资源池ID
        - name: queueId
          in: query
          schema:
            type: string
          description: 队列ID
        - name: nodeName
          in: query
          schema:
            type: string
          description: 节点名称
        - name: type
          in: query
          schema:
            type: string
            enum:
              - selfManaged
              - managed
            default: selfManaged
        - name: keyword
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueDetail"
  /aihc/resource-pools/DescribeNodes:
    get:
      tags:
        - ResourcePools
      summary: ResourcePools - DescribeNodes
      operationId: describeNodes
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: resourcePoolId
          in: query
          schema:
            type: string
          description: 资源池ID
        - name: queueId
          in: query
          schema:
            type: string
          description: 队列ID
        - name: nodeName
          in: query
          schema:
            type: string
          description: 节点名称
        - name: type
          in: query
          schema:
            type: string
            enum:
              - selfManaged
              - managed
            default: selfManaged
        - name: keyword
          in: query
          schema:
            type: string
        - name: pageNumber
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeListResponse"
  /aihc/resource-pools/DescribeNode:
    get:
      tags:
        - ResourcePools
      summary: ResourcePools - DescribeNode
      operationId: describeNode
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: resourcePoolId
          in: query
          schema:
            type: string
          description: 资源池ID
        - name: queueId
          in: query
          schema:
            type: string
          description: 队列ID
        - name: nodeName
          in: query
          schema:
            type: string
          description: 节点名称
        - name: type
          in: query
          schema:
            type: string
            enum:
              - selfManaged
              - managed
            default: selfManaged
        - name: keyword
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeDetail"
  /aihc/resource-pools/DescribeResourcePoolUsage:
    get:
      tags:
        - ResourcePools
      summary: ResourcePools - DescribeResourcePoolUsage
      operationId: describeResourcePoolUsage
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
        - name: resourcePoolId
          in: query
          schema:
            type: string
          description: 资源池ID
        - name: queueId
          in: query
          schema:
            type: string
          description: 队列ID
        - name: nodeName
          in: query
          schema:
            type: string
          description: 节点名称
        - name: type
          in: query
          schema:
            type: string
            enum:
              - selfManaged
              - managed
            default: selfManaged
        - name: keyword
          in: query
          schema:
            type: string
        - name: pageNumber
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourcePoolUsage"
...
  /aihc/jobs/StopJob:
    post:
      tags:
        - Jobs
      summary: Jobs - StopJob
      operationId: stopJob
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: 动态透传参数
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /aihc/jobs/BatchStopJobs:
    post:
      tags:
        - Jobs
      summary: Jobs - BatchStopJobs
      operationId: batchStopJobs
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: 动态透传参数
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /aihc/jobs/ModifyJob:
    post:
      tags:
        - Jobs
      summary: Jobs - ModifyJob
      operationId: modifyJob
      parameters:
        - $ref: "#/components/parameters/RegionParam"
        - $ref: "#/components/parameters/AkParam"
        - $ref: "#/components/parameters/SkParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: 动态透传参数
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
components:
  parameters:
    RegionParam:
      name: region
      in: header
      required: true
      description: 区域代码
      schema:
        type: string
        enum:
          - bj
          - gz
          - su
          - bd
          - fwh
          - yq
        default: bd
    AkParam:
      name: ak
      in: header
      required: true
      description: 百度云 Access Key
      schema:
        type: string
    SkParam:
      name: sk
      in: header
      required: true
      description: 百度云 Secret Key
      schema:
        type: string
  schemas:
...
    CreateJobResponse:
      type: object
      properties:
        requestId:
          type: string
        jobId:
          type: string
        jobName:
          type: string
    DeleteJobResponse:
      type: object
      properties:
        requestId:
          type: string
        jobId:
          type: string
    JobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            type: object
        pagination:
          $ref: "#/components/schemas/Pagination"
    JobDetail:
      type: object
      properties:
        jobId:
          type: string
        jobName:
          type: string
        status:
          type: string
        resourcePool:
          $ref: "#/components/schemas/ResourcePoolRef"
        framework:
          type: string
        command:
          type: string
        resources:
          $ref: "#/components/schemas/ComputeResources"
        replicas:
          type: integer
        image:
          $ref: "#/components/schemas/ImageConfig"
        envs:
          type: object
          additionalProperties:
            type: string
        labels:
          type: object
          additionalProperties:
            type: string
        priority:
          type: string
        enableRDMA:
          type: boolean
...